+++ zebROS_ws/src/as726x_controllers/as726x_controllers_plugin.xml
@@ -4,6 +4,11 @@
 			The as726x state controller publishes the current as726x state as an frc_msgs/AS726xState message. The as726x state controller expects a as726x::AS726xStateInterface type of hardware interface.
 		</description>
 	</class>
+	<class name="state_listener_controller/AS726xStateListenerController" type="as726x_state_controller::AS726xStateListenerController" base_class_type="controller_interface::ControllerBase">
+		<description>
+			Listens to as726x::AS726xState messages for imus marked as remote, writes their value to the local controller.
+		</description>
+	</class>
 	<class name="as726x_controller/AS726xController" type="as726x_controller::AS726xController" base_class_type="controller_interface::ControllerBase">
 		<description>
 			The as726x controller allows configuration of sensor setting through either params or dynamic reconfigure
