+++ zebROS_ws/src/frc_state_controllers/include/frc_state_controllers/match_state_controller.h
@@ -1,6 +1,7 @@
 #pragma once
 
 #include <controller_interface/controller.h>
+#include <realtime_tools/realtime_buffer.h>
 #include <realtime_tools/realtime_publisher.h>
 #include <frc_interfaces/match_data_interface.h>
 #include "frc_msgs/MatchSpecificData.h"
@@ -12,13 +13,13 @@ class MatchStateController: public controller_interface::Controller<hardware_int
 	public:
 		MatchStateController() : publish_rate_(15.0) {}
 
-		virtual bool init(hardware_interface::MatchStateInterface *hw,
-						ros::NodeHandle					&root_nh,
-						ros::NodeHandle					&controller_nh) override;
+		bool init(hardware_interface::MatchStateInterface *hw,
+				  ros::NodeHandle						  &root_nh,
+				  ros::NodeHandle						  &controller_nh) override;
 
-		virtual void starting(const ros::Time &time) override;
-		virtual void update(const ros::Time &time, const ros::Duration & ) override;
-		virtual void stopping(const ros::Time &time) override;
+		void starting(const ros::Time &time) override;
+		void update(const ros::Time &time, const ros::Duration & ) override;
+		void stopping(const ros::Time &time) override;
 
 	private:
 		hardware_interface::MatchStateHandle match_state_;
@@ -28,4 +29,32 @@ class MatchStateController: public controller_interface::Controller<hardware_int
 
 }; //class
 
-} //namespace
+} //namespace match_state_controller
+
+namespace state_listener_controller
+{
+class MatchStateListenerController :
+	public controller_interface::Controller<hardware_interface::RemoteMatchStateInterface>
+{
+	public:
+		MatchStateListenerController();
+		~MatchStateListenerController();
+
+		bool init(hardware_interface::RemoteMatchStateInterface *hw, ros::NodeHandle &n) override;
+		void starting(const ros::Time & /*time*/) override;
+		void stopping(const ros::Time & /*time*/) override;
+		void update(const ros::Time & /*time*/, const ros::Duration & /*period*/) override;
+
+	private:
+		ros::Subscriber sub_command_;
+		hardware_interface::MatchStateWritableHandle handle_;
+
+		// Real-time buffer holds the last command value read from the "command" topic.
+		realtime_tools::RealtimeBuffer<hardware_interface::MatchHWState> command_buffer_;
+
+		// Iterate through each desired joint state.  If it is found in
+		// the message, save the value here in the realtime buffer.
+		void commandCB(const frc_msgs::MatchSpecificDataConstPtr &msg);
+};
+
+} // namespace state_listener_controller
