+++ zebROS_ws/src/frc_state_controllers/include/frc_state_controllers/pdp_state_controller.h
@@ -2,6 +2,7 @@
 
 #include <controller_interface/controller.h>
 #include <hardware_interface/joint_state_interface.h>
+#include <realtime_tools/realtime_buffer.h>
 #include <realtime_tools/realtime_publisher.h>
 #include <frc_msgs/PDPData.h>
 #include <frc_interfaces/pdp_state_interface.h>
@@ -13,15 +14,16 @@ namespace pdp_state_controller
 class PDPStateController: public controller_interface::Controller<hardware_interface::PDPStateInterface>
 {
 	public:
-		PDPStateController() : publish_rate_(15.0) {
-                }
+		PDPStateController() : publish_rate_(15.0)
+		{
+        }
 
-		virtual bool init(hardware_interface::PDPStateInterface *hw,
-							ros::NodeHandle						&root_nh,
-							ros::NodeHandle						&controller_nh) override;
-		virtual void starting(const ros::Time &time) override;
-		virtual void update(const ros::Time &time, const ros::Duration & ) override;
-		virtual void stopping(const ros::Time &time) override;
+		bool init(hardware_interface::PDPStateInterface *hw,
+				  ros::NodeHandle						&root_nh,
+				  ros::NodeHandle						&controller_nh) override;
+		void starting(const ros::Time &time) override;
+		void update(const ros::Time &time, const ros::Duration & ) override;
+		void stopping(const ros::Time &time) override;
 
 	private:
 		hardware_interface::PDPStateHandle pdp_state_;
@@ -30,5 +32,31 @@ class PDPStateController: public controller_interface::Controller<hardware_inter
 		double publish_rate_;
 
 }; //class
+}
 
+namespace state_listener_controller
+{
+class PDPStateListenerController :
+	public controller_interface::Controller<hardware_interface::RemotePDPStateInterface>
+{
+	public:
+		PDPStateListenerController();
+		~PDPStateListenerController();
+
+		bool init(hardware_interface::RemotePDPStateInterface *hw, ros::NodeHandle &n) override;
+		void starting(const ros::Time & /*time*/) override;
+		void stopping(const ros::Time & /*time*/) override;
+		void update(const ros::Time & /*time*/, const ros::Duration & /*period*/) override;
+
+	private:
+		ros::Subscriber sub_command_;
+		hardware_interface::PDPWritableStateHandle handle_;
+
+		// Real-time buffer holds the last command value read from the "command" topic.
+		realtime_tools::RealtimeBuffer<hardware_interface::PDPHWState> command_buffer_;
+
+		// Iterate through each desired joint state.  If it is found in
+		// the message, save the value here in the realtime buffer.
+		virtual void commandCB(const frc_msgs::PDPDataConstPtr &msg);
+};
 } //namespace
