+++ zebROS_ws/src/remote_joint_interface/README.md
@@ -1,0 +1,3 @@
+Used to create special writable joint state interfaces. Used by
+state_listener to mirror joint state between remote and local
+hardware interfaces
