+++ zebROS_ws/src/remote_joint_interface/include/remote_joint_interface/remote_joint_interface.h
@@ -5,9 +5,6 @@
 #define private protected
 #include <hardware_interface/imu_sensor_interface.h>
 #undef private
-#include "frc_interfaces/match_data_interface.h"
-#include "frc_interfaces/pcm_state_interface.h"
-#include "frc_interfaces/pdp_state_interface.h"
 
 // Create a separate type of joint interface for joints which
 // are read from hardware on another controller_manager.  This
@@ -59,7 +56,4 @@ namespace hardware_interface
 	};
 
 	class RemoteImuSensorInterface  : public HardwareResourceManager<ImuWritableSensorHandle,  ClaimResources> {};
-	class RemotePDPStateInterface   : public HardwareResourceManager<PDPWritableStateHandle,   ClaimResources> {};
-	class RemotePCMStateInterface   : public HardwareResourceManager<PCMWritableStateHandle,   ClaimResources> {};
-	class RemoteMatchStateInterface : public HardwareResourceManager<MatchStateWritableHandle, ClaimResources> {};
 }
