+++ zebROS_ws/src/remote_joint_interface/package.xml
@@ -1,13 +1,13 @@
 <?xml version="1.0"?>
 <package format="2">
-  <name>state_listener</name>
+  <name>remote_joint_interface</name>
   <version>0.0.0</version>
-  <description>The state_listener package</description>
+  <description>The remote_joint_interface package</description>
 
   <!-- One maintainer tag required, multiple allowed, one person per tag -->
   <!-- Example:  -->
   <!-- <maintainer email="jane.doe@example.com">Jane Doe</maintainer> -->
-  <maintainer email="programmers@team900.org">Team 900 Programmers</maintainer>
+  <maintainer email="kjaget@todo.todo">kjaget</maintainer>
 
 
   <!-- One license tag required, multiple allowed, one license per tag -->
@@ -19,7 +19,7 @@
   <!-- Url tags are optional, but multiple are allowed, one per tag -->
   <!-- Optional attribute type can be: website, bugtracker, or repository -->
   <!-- Example: -->
-  <!-- <url type="website">http://wiki.ros.org/joint_state_listener</url> -->
+  <!-- <url type="website">http://wiki.ros.org/remote_joint_interface</url> -->
 
 
   <!-- Author tags are optional, multiple are allowed, one per tag -->
@@ -49,26 +49,14 @@
   <!-- Use doc_depend for packages you need only for building documentation: -->
   <!--   <doc_depend>doxygen</doc_depend> -->
   <buildtool_depend>catkin</buildtool_depend>
-  <build_depend>controller_interface</build_depend>
-  <build_depend>realtime_tools</build_depend>
   <build_depend>hardware_interface</build_depend>
-  <build_depend>talon_interface</build_depend>
-  <build_depend>talon_state_msgs</build_depend>
-  <build_depend>roscpp</build_depend>
-  <build_export_depend>controller_interface</build_export_depend>
-  <build_export_depend>realtime_tools</build_export_depend>
-  <build_export_depend>roscpp</build_export_depend>
-
-  <exec_depend>controller_interface</exec_depend>
-  <exec_depend>realtime_tools</exec_depend>
+  <build_export_depend>hardware_interface</build_export_depend>
   <exec_depend>hardware_interface</exec_depend>
-  <exec_depend>roscpp</exec_depend>
-  <depend>frc_msgs</depend>
-  <depend>remote_joint_interface</depend>
+
 
   <!-- The export tag contains other, unspecified, tags -->
   <export>
     <!-- Other tools can request additional information be placed here -->
-    <controller_interface plugin="${prefix}/state_listener_plugins.xml"/>
+
   </export>
 </package>
