+++ zebROS_ws/src/ros_control_boilerplate/include/ros_control_boilerplate/frc_robot_interface.h
@@ -52,8 +52,8 @@
 #include "frc_interfaces/pcm_state_interface.h"
 #include "frc_interfaces/pdp_state_interface.h"
 #include "frc_interfaces/match_data_interface.h"
-#include "frc_interfaces/remote_joint_interface.h"
 #include "frc_interfaces/robot_controller_interface.h"
+#include "remote_joint_interface/remote_joint_interface.h"
 #include "talon_interface/cancoder_command_interface.h"
 #include "talon_interface/canifier_command_interface.h"
 #include "talon_interface/talon_command_interface.h"
