+++ zebROS_ws/src/state_listener/README.md
@@ -1,0 +1,4 @@
+Code to create state listener controllers for generic ROS joint types
+These controllers are used to monitor joint_state controllers on remote
+hardware interfaces and copy the state into the internal buffers on the
+local hardware interface.
