+++ zebROS_ws/src/state_listener/include/joint_state_listener/state_listener_controller.h
@@ -5,14 +5,8 @@
 #include <realtime_tools/realtime_buffer.h>
 #include <sensor_msgs/JointState.h>
 #include <sensor_msgs/Imu.h>
-#include "as726x_interface/as726x_interface.h"
-#include "as726x_msgs/AS726xState.h"
-#include "frc_interfaces/remote_joint_interface.h"
 #include "frc_msgs/JointMode.h"
-#include "frc_msgs/MatchSpecificData.h"
-#include "frc_msgs/PDPData.h"
-#include "talon_interface/talon_state_interface.h"
-#include "talon_state_msgs/TalonState.h"
+#include "remote_joint_interface/remote_joint_interface.h"
 
 namespace state_listener_controller
 {
@@ -57,7 +51,7 @@ class JointStateListenerController :
 			sub_command_.shutdown();
 		}
 
-		virtual bool init(hardware_interface::RemoteJointInterface *hw, ros::NodeHandle &n) override
+		bool init(hardware_interface::RemoteJointInterface *hw, ros::NodeHandle &n) override
 		{
 			// Read list of hw, make a list, grab handles for them, plus allocate storage space
 			joint_names_ = hw->getNames();
@@ -82,15 +76,15 @@ class JointStateListenerController :
 			return true;
 		}
 
-		virtual void starting(const ros::Time & /*time*/) override
+		void starting(const ros::Time & /*time*/) override
 		{
 		}
-		virtual void stopping(const ros::Time & /*time*/) override
+		void stopping(const ros::Time & /*time*/) override
 		{
 			//handles_.release();
 		}
 
-		virtual void update(const ros::Time & /*time*/, const ros::Duration & /*period*/) override
+		void update(const ros::Time & /*time*/, const ros::Duration & /*period*/) override
 		{
 			// Take the most recent set of values read from the joint_states
 			// topic and write them to the local joints
@@ -110,7 +104,7 @@ class JointStateListenerController :
 
 		// Iterate through each desired joint state.  If it is found in
 		// the message, save the value here in the realtime buffer.
-		virtual void commandCB(const sensor_msgs::JointStateConstPtr &msg)
+		void commandCB(const sensor_msgs::JointStateConstPtr &msg)
 		{
 			std::vector<ValueValid<double>> ret;
 			ret.resize(joint_names_.size());
@@ -138,7 +132,7 @@ class JointModeListenerController :
 			sub_command_.shutdown();
 		}
 
-		virtual bool init(hardware_interface::RemoteJointModeInterface *hw, ros::NodeHandle &n) override
+		bool init(hardware_interface::RemoteJointModeInterface *hw, ros::NodeHandle &n) override
 		{
 			// Read list of hw, make a list, grab handles for them, plus allocate storage space
 			joint_names_ = hw->getNames();
@@ -163,14 +157,14 @@ class JointModeListenerController :
 			return true;
 		}
 
-		virtual void starting(const ros::Time & /*time*/) override
+		void starting(const ros::Time & /*time*/) override
 		{
 		}
-		virtual void stopping(const ros::Time & /*time*/) override
+		void stopping(const ros::Time & /*time*/) override
 		{
 		}
 
-		virtual void update(const ros::Time & /*time*/, const ros::Duration & /*period*/) override
+		void update(const ros::Time & /*time*/, const ros::Duration & /*period*/) override
 		{
 			// Take the most recent set of values read from the joint_modes
 			// topic and write them to the local joints
@@ -190,7 +184,7 @@ class JointModeListenerController :
 
 		// Iterate through each desired joint mode.  If it is found in
 		// the message, save the value here in the realtime buffer.
-		virtual void commandCB(const frc_msgs::JointModeConstPtr &msg)
+		void commandCB(const frc_msgs::JointModeConstPtr &msg)
 		{
 			std::vector<ValueValid<int>> ret;
 			ret.resize(joint_names_.size());
@@ -208,169 +202,6 @@ class JointModeListenerController :
 		}
 };
 
-
-class PDPStateListenerController :
-	public controller_interface::Controller<hardware_interface::RemotePDPStateInterface>
-{
-	public:
-		PDPStateListenerController() {}
-		~PDPStateListenerController()
-		{
-			sub_command_.shutdown();
-		}
-
-		virtual bool init(hardware_interface::RemotePDPStateInterface *hw, ros::NodeHandle &n) override
-		{
-			// Read list of hw, make a list, grab handles for them, plus allocate storage space
-			auto joint_names = hw->getNames();
-			if (joint_names.size() == 0)
-			{
-				ROS_ERROR("PDP State Listener Controller : no remote pdp joints defined");
-			}
-			ROS_INFO_STREAM("PDP State Listener Controller got joint " << joint_names[0]);
-			handle_ = hw->getHandle(joint_names[0]);
-
-			std::string topic;
-
-			// get topic to subscribe to
-			if (!n.getParam("topic", topic))
-			{
-				ROS_ERROR("Parameter 'topic' not set");
-				return false;
-			}
-
-			sub_command_ = n.subscribe<frc_msgs::PDPData>(topic, 1, &PDPStateListenerController::commandCB, this);
-			return true;
-		}
-
-		virtual void starting(const ros::Time & /*time*/) override
-		{
-		}
-		virtual void stopping(const ros::Time & /*time*/) override
-		{
-		}
-
-		virtual void update(const ros::Time & /*time*/, const ros::Duration & /*period*/) override
-		{
-			// Quick way to do a shallow copy of the entire HW state
-			*(handle_.operator->()) = *command_buffer_.readFromRT();
-		}
-
-	private:
-		ros::Subscriber sub_command_;
-		hardware_interface::PDPWritableStateHandle handle_;
-
-		// Real-time buffer holds the last command value read from the
-		// "command" topic.
-		realtime_tools::RealtimeBuffer<hardware_interface::PDPHWState> command_buffer_;
-
-		// Iterate through each desired joint state.  If it is found in
-		// the message, save the value here in the realtime buffer.
-		// // TODO : figure out how to hack this to use a ConstPtr type instead
-		virtual void commandCB(const frc_msgs::PDPDataConstPtr &msg)
-		{
-			hardware_interface::PDPHWState data;
-
-			data.setVoltage(msg->voltage);
-			data.setTemperature(msg->temperature);
-			data.setTotalCurrent(msg->totalCurrent);
-			data.setTotalPower(msg->totalPower);
-			data.setTotalEnergy(msg->totalEnergy);
-			for (size_t channel = 0; channel <= 15; channel++)
-				data.setCurrent(msg->current[channel], channel);
-			command_buffer_.writeFromNonRT(data);
-		}
-};
-
-class MatchStateListenerController :
-	public controller_interface::Controller<hardware_interface::RemoteMatchStateInterface>
-{
-	public:
-		MatchStateListenerController() {}
-		~MatchStateListenerController()
-		{
-			sub_command_.shutdown();
-		}
-
-		virtual bool init(hardware_interface::RemoteMatchStateInterface *hw, ros::NodeHandle &n) override
-		{
-			// Read list of hw, make a list, grab handles for them, plus allocate storage space
-			auto joint_names = hw->getNames();
-			if (joint_names.size() == 0)
-			{
-				ROS_ERROR("Match State Listener Controller : no remote match joints defined. Don't run this on a roboRio");
-			}
-			ROS_INFO_STREAM("Match State Listener Controller got joint " << joint_names[0]);
-			handle_ = hw->getHandle(joint_names[0]);
-
-			std::string topic;
-
-			// get topic to subscribe to
-			if (!n.getParam("topic", topic))
-			{
-				ROS_ERROR("Parameter 'topic' not set");
-				return false;
-			}
-
-			sub_command_ = n.subscribe<frc_msgs::MatchSpecificData>(topic, 1, &MatchStateListenerController::commandCB, this);
-			return true;
-		}
-
-		virtual void starting(const ros::Time & /*time*/) override
-		{
-		}
-		virtual void stopping(const ros::Time & /*time*/) override
-		{
-		}
-
-		virtual void update(const ros::Time & /*time*/, const ros::Duration & /*period*/) override
-		{
-			// Quick way to do a shallow copy of the entire HW state
-			*(handle_.operator->()) = *command_buffer_.readFromRT();
-		}
-
-	private:
-		ros::Subscriber sub_command_;
-		hardware_interface::MatchStateWritableHandle handle_;
-
-		// Real-time buffer holds the last command value read from the
-		// "command" topic.
-		realtime_tools::RealtimeBuffer<hardware_interface::MatchHWState> command_buffer_;
-
-		// Iterate through each desired joint state.  If it is found in
-		// the message, save the value here in the realtime buffer.
-		// // TODO : figure out how to hack this to use a ConstPtr type instead
-		virtual void commandCB(const frc_msgs::MatchSpecificDataConstPtr &msg)
-		{
-			hardware_interface::MatchHWState data;
-			data.setMatchTimeRemaining(msg->matchTimeRemaining);
-
-			data.setGameSpecificData(msg->gameSpecificData);
-			data.setEventName(msg->eventName);
-
-			data.setAllianceColor(msg->allianceColor);
-			data.setMatchType(msg->matchType);
-			data.setDriverStationLocation(msg->driverStationLocation);
-			data.setMatchNumber(msg->matchNumber);
-			data.setReplayNumber(msg->replayNumber);
-
-			data.setEnabled(msg->Enabled);
-			data.setDisabled(msg->Disabled);
-			data.setAutonomous(msg->Autonomous);
-			data.setFMSAttached(msg->FMSAttached);
-			data.setDSAttached(msg->DSAttached);
-			data.setOperatorControl(msg->OperatorControl);
-			data.setTest(msg->Test);
-
-			data.setBatteryVoltage(msg->BatteryVoltage);
-
-			data.setGetMatchTimeStatus(msg->getMatchTimeStatus);
-			data.setGetAllianceStationStatus(msg->getAllianceStationStatus);
-			data.setGetVinVoltageStatus(msg->getVinVoltageStatus);
-			command_buffer_.writeFromNonRT(data);
-		}
-};
-
 class IMUStateListenerController :
 	public controller_interface::Controller<hardware_interface::RemoteImuSensorInterface>
 {
@@ -381,7 +212,7 @@ class IMUStateListenerController :
 			sub_command_.shutdown();
 		}
 
-		virtual bool init(hardware_interface::RemoteImuSensorInterface *hw, ros::NodeHandle &n) override
+		bool init(hardware_interface::RemoteImuSensorInterface *hw, ros::NodeHandle &n) override
 		{
 			// Read list of hw, make a list, grab handles for them, plus allocate storage space
 			auto joint_names = hw->getNames();
@@ -405,14 +236,14 @@ class IMUStateListenerController :
 			return true;
 		}
 
-		virtual void starting(const ros::Time & /*time*/) override
+		void starting(const ros::Time & /*time*/) override
 		{
 		}
-		virtual void stopping(const ros::Time & /*time*/) override
+		void stopping(const ros::Time & /*time*/) override
 		{
 		}
 
-		virtual void update(const ros::Time & /*time*/, const ros::Duration & /*period*/) override
+		void update(const ros::Time & /*time*/, const ros::Duration & /*period*/) override
 		{
 			const auto data = *command_buffer_.readFromRT();
 
@@ -439,7 +270,7 @@ class IMUStateListenerController :
 		std::array<double, 3> data_linear_acceleration_;
 		std::array<double, 9> data_linear_acceleration_covariance_;
 
-		virtual void commandCB(const sensor_msgs::ImuConstPtr &msg)
+		void commandCB(const sensor_msgs::ImuConstPtr &msg)
 		{
 			hardware_interface::ImuSensorHandle::Data data;
 
@@ -470,273 +301,4 @@ class IMUStateListenerController :
 		}
 };
 
-class TalonStateListenerController :
-	public controller_interface::Controller<hardware_interface::RemoteTalonStateInterface>
-{
-	public:
-		TalonStateListenerController() {}
-		~TalonStateListenerController()
-		{
-			sub_command_.shutdown();
-		}
-
-		virtual bool init(hardware_interface::RemoteTalonStateInterface *hw, ros::NodeHandle &n) override
-		{
-			// Read list of hw, make a list, grab handles for them, plus allocate storage space
-			joint_names_ = hw->getNames();
-			for (auto j : joint_names_)
-			{
-				ROS_INFO_STREAM("Joint State Listener Controller got joint " << j);
-				handles_.push_back(hw->getHandle(j));
-			}
-
-			std::string topic;
-
-			// get topic to subscribe to
-			if (!n.getParam("topic", topic))
-			{
-				ROS_ERROR("Parameter 'topic' not set");
-				return false;
-			}
-
-			sub_command_ = n.subscribe<talon_state_msgs::TalonState>(topic, 1, &TalonStateListenerController::commandCB, this);
-			return true;
-		}
-
-		virtual void starting(const ros::Time & /*time*/) override
-		{
-		}
-		virtual void stopping(const ros::Time & /*time*/) override
-		{
-			//handles_.release();
-		}
-
-		virtual void update(const ros::Time & /*time*/, const ros::Duration & /*period*/) override
-		{
-			// Take the most recent set of values read from the joint_states
-			// topic and write them to the local joints
-			auto vals = *command_buffer_.readFromRT();
-			for (size_t i = 0; i < vals.size(); i++)
-			{
-				if (vals[i].valid_)
-				{
-					auto ts = vals[i].value_;
-					handles_[i]->setPosition(ts.getPosition());
-					handles_[i]->setSpeed(ts.getSpeed());
-					handles_[i]->setOutputCurrent(ts.getOutputCurrent());
-					handles_[i]->setBusVoltage(ts.getBusVoltage());
-					handles_[i]->setMotorOutputPercent(ts.getMotorOutputPercent());
-					handles_[i]->setOutputVoltage(ts.getOutputVoltage());
-					handles_[i]->setTemperature(ts.getTemperature());
-					handles_[i]->setClosedLoopError(ts.getClosedLoopError());
-					handles_[i]->setIntegralAccumulator(ts.getIntegralAccumulator());
-					handles_[i]->setErrorDerivative(ts.getErrorDerivative());
-					handles_[i]->setClosedLoopTarget(ts.getClosedLoopTarget());
-					handles_[i]->setActiveTrajectoryPosition(ts.getActiveTrajectoryPosition());
-					handles_[i]->setActiveTrajectoryVelocity(ts.getActiveTrajectoryVelocity());
-					handles_[i]->setActiveTrajectoryHeading(ts.getActiveTrajectoryHeading());
-					handles_[i]->setMotionProfileTopLevelBufferCount(ts.getMotionProfileTopLevelBufferCount());
-					handles_[i]->setFaults(ts.getFaults());
-					handles_[i]->setForwardLimitSwitch(ts.getForwardLimitSwitch());
-					handles_[i]->setReverseLimitSwitch(ts.getReverseLimitSwitch());
-					handles_[i]->setForwardSoftlimitHit(ts.getForwardSoftlimitHit());
-					handles_[i]->setReverseSoftlimitHit(ts.getReverseSoftlimitHit());
-					handles_[i]->setStickyFaults(ts.getStickyFaults());
-				}
-			}
-
-		}
-
-	private:
-		ros::Subscriber sub_command_;
-		std::vector<std::string> joint_names_;
-		std::vector<hardware_interface::TalonWritableStateHandle> handles_;
-
-		// Real-time buffer holds the last command value read from the
-		// "command" topic.
-		realtime_tools::RealtimeBuffer<std::vector<ValueValid<hardware_interface::TalonHWState>>> command_buffer_;
-
-		virtual void commandCB(const talon_state_msgs::TalonStateConstPtr &msg)
-		{
-			std::vector<ValueValid<hardware_interface::TalonHWState>> data;
-			for (size_t i = 0; i < joint_names_.size(); i++)
-				data.push_back(hardware_interface::TalonHWState(0)); // dummy can ID since it isn't used
-			for (size_t i = 0; i < joint_names_.size(); i++)
-			{
-				auto it = std::find(msg->name.cbegin(), msg->name.cend(), joint_names_[i]);
-				if (it != msg->name.cend())
-				{
-					data[i].value_.setPosition(msg->position[i]);
-					data[i].value_.setSpeed(msg->speed[i]);
-					data[i].value_.setOutputCurrent(msg->output_voltage[i]);
-					data[i].value_.setBusVoltage(msg->bus_voltage[i]);
-					data[i].value_.setMotorOutputPercent(msg->motor_output_percent[i]);
-					data[i].value_.setOutputVoltage(msg->output_voltage[i]);
-					data[i].value_.setTemperature(msg->temperature[i]);
-					data[i].value_.setClosedLoopError(msg->closed_loop_error[i]);
-					data[i].value_.setIntegralAccumulator(msg->integral_accumulator[i]);
-					data[i].value_.setErrorDerivative(msg->error_derivative[i]);
-					data[i].value_.setClosedLoopTarget(msg->closed_loop_target[i]);
-					data[i].value_.setActiveTrajectoryPosition(msg->active_trajectory_position[i]);
-					data[i].value_.setActiveTrajectoryVelocity(msg->active_trajectory_velocity[i]);
-					data[i].value_.setActiveTrajectoryHeading(msg->active_trajectory_heading[i]);
-					data[i].value_.setMotionProfileTopLevelBufferCount(msg->motion_profile_top_level_buffer_count[i]);
-					//data[i].value_.setFaults(msg->getFaults[i]);
-					data[i].value_.setForwardLimitSwitch(msg->forward_limit_switch[i]);
-					data[i].value_.setReverseLimitSwitch(msg->reverse_limit_switch[i]);
-					data[i].value_.setForwardSoftlimitHit(msg->forward_softlimit[i]);
-					data[i].value_.setReverseSoftlimitHit(msg->reverse_softlimit[i]);
-					//data[i].value_.setStickyFaults(msg->getStickyFaults[i]);
-					data[i].valid_ = true;
-				}
-			}
-			command_buffer_.writeFromNonRT(data);
-		}
-};
-
-class AS726xStateListenerController :
-	public controller_interface::Controller<hardware_interface::as726x::RemoteAS726xStateInterface>
-{
-	public:
-		AS726xStateListenerController() {}
-		~AS726xStateListenerController()
-		{
-			sub_command_.shutdown();
-		}
-
-		virtual bool init(hardware_interface::as726x::RemoteAS726xStateInterface *hw, ros::NodeHandle &n) override
-		{
-			// Read list of hw, make a list, grab handles for them, plus allocate storage space
-			joint_names_ = hw->getNames();
-			if (joint_names_.size() == 0)
-			{
-				ROS_ERROR("AS726x State Listener Controller : no remote as726x joints defined");
-			}
-			for (auto j : joint_names_)
-			{
-				ROS_INFO_STREAM("AS726x State Listener Controller got joint " << j);
-				handles_.push_back(hw->getHandle(j));
-			}
-
-			std::string topic;
-
-			// get topic to subscribe to
-			if (!n.getParam("topic", topic))
-			{
-				ROS_ERROR("Parameter 'topic' not set");
-				return false;
-			}
-
-			sub_command_ = n.subscribe<as726x_msgs::AS726xState>(topic, 1, &AS726xStateListenerController::commandCB, this);
-			return true;
-		}
-
-		virtual void starting(const ros::Time & /*time*/) override
-		{
-		}
-		virtual void stopping(const ros::Time & /*time*/) override
-		{
-			//handles_.release();
-		}
-
-		virtual void update(const ros::Time & /*time*/, const ros::Duration & /*period*/) override
-		{
-			// Take the most recent set of values read from the joint_states
-			// topic and write them to the local joints
-			auto vals = *command_buffer_.readFromRT();
-			for (size_t i = 0; i < vals.size(); i++)
-				if (vals[i].valid_)
-					// Quick way to do a shallow copy of the entire HW state
-					*(handles_[i].operator->()) = vals[i].value_;
-		}
-
-	private:
-		ros::Subscriber sub_command_;
-		std::vector<std::string> joint_names_;
-		std::vector<hardware_interface::as726x::AS726xWritableStateHandle> handles_;
-
-		// Real-time buffer holds the last command value read from the
-		// "command" topic.
-		realtime_tools::RealtimeBuffer<std::vector<ValueValid<hardware_interface::as726x::AS726xState>>> command_buffer_;
-
-		// Iterate through each desired joint state.  If it is found in
-		// the message, save the value here in the realtime buffer.
-		// // TODO : figure out how to hack this to use a ConstPtr type instead
-		virtual void commandCB(const as726x_msgs::AS726xStateConstPtr &msg)
-		{
-			std::vector<ValueValid<hardware_interface::as726x::AS726xState>> data;
-			for (size_t i = 0; i < joint_names_.size(); i++)
-			{
-				auto it = std::find(msg->name.cbegin(), msg->name.cend(), joint_names_[i]);
-				if (it != msg->name.cend())
-				{
-					const size_t loc = it - msg->name.cbegin();
-					data.push_back(hardware_interface::as726x::AS726xState(msg->port[loc], msg->address[loc]));
-
-					hardware_interface::as726x::IndLedCurrentLimits ind_led_current_limit = static_cast< hardware_interface::as726x::IndLedCurrentLimits>(0);
-					if (msg->ind_led_current_limit[loc] == "1mA")
-						ind_led_current_limit = hardware_interface::as726x::IndLedCurrentLimits::IND_LIMIT_1MA;
-					else if (msg->ind_led_current_limit[loc] == "2mA")
-						ind_led_current_limit = hardware_interface::as726x::IndLedCurrentLimits::IND_LIMIT_2MA;
-					else if (msg->ind_led_current_limit[loc] == "4mA")
-						ind_led_current_limit = hardware_interface::as726x::IndLedCurrentLimits::IND_LIMIT_4MA;
-					else if (msg->ind_led_current_limit[loc] == "8mA")
-						ind_led_current_limit = hardware_interface::as726x::IndLedCurrentLimits::IND_LIMIT_8MA;
-					data[i].value_.setIndLedCurrentLimit(ind_led_current_limit);
-					data[i].value_.setIndLedEnable(msg->ind_led_enable[loc]);
-
-					hardware_interface::as726x::DrvLedCurrentLimits drv_led_current_limit = static_cast<hardware_interface::as726x::DrvLedCurrentLimits>(0);
-					if (msg->drv_led_current_limit[loc] == "12mA5")
-						drv_led_current_limit = hardware_interface::as726x::DrvLedCurrentLimits::DRV_LIMIT_12MA5;
-					else if (msg->drv_led_current_limit[loc] == "25mA")
-						drv_led_current_limit = hardware_interface::as726x::DrvLedCurrentLimits::DRV_LIMIT_25MA;
-					else if (msg->drv_led_current_limit[loc] == "50mA")
-						drv_led_current_limit = hardware_interface::as726x::DrvLedCurrentLimits::DRV_LIMIT_50MA;
-					else if (msg->drv_led_current_limit[loc] == "100mA")
-						drv_led_current_limit = hardware_interface::as726x::DrvLedCurrentLimits::DRV_LIMIT_100MA;
-					data[i].value_.setDrvLedCurrentLimit(drv_led_current_limit);
-
-					hardware_interface::as726x::ConversionTypes conversion_type = static_cast<hardware_interface::as726x::ConversionTypes>(0);
-					if (msg->conversion_type[loc] == "MODE_0")
-						conversion_type = hardware_interface::as726x::ConversionTypes::MODE_0;
-					else if (msg->conversion_type[loc] == "MODE_1")
-						conversion_type = hardware_interface::as726x::ConversionTypes::MODE_1;
-					else if (msg->conversion_type[loc] == "MODE_2")
-						conversion_type = hardware_interface::as726x::ConversionTypes::MODE_2;
-					else if (msg->conversion_type[loc] == "ONE_SHOT")
-						conversion_type = hardware_interface::as726x::ConversionTypes::ONE_SHOT;
-					data[i].value_.setConversionType(conversion_type);
-
-					hardware_interface::as726x::ChannelGain gain = static_cast<hardware_interface::as726x::ChannelGain>(0);
-					if (msg->gain[loc] == "GAIN_1X")
-						gain = hardware_interface::as726x::ChannelGain::GAIN_1X;
-					else if (msg->gain[loc] == "GAIN_3X7")
-						gain = hardware_interface::as726x::ChannelGain::GAIN_3X7;
-					else if (msg->gain[loc] == "GAIN_16X")
-						gain = hardware_interface::as726x::ChannelGain::GAIN_16X;
-					else if (msg->gain[loc] == "GAIN_64X")
-						gain = hardware_interface::as726x::ChannelGain::GAIN_64X;
-					data[i].value_.setGain(gain);
-					data[i].value_.setIntegrationTime(msg->integration_time[loc]);
-					data[i].value_.setTemperature(msg->temperature[loc]);
-					std::array<uint16_t, 6> raw_channel_data;
-					std::array<float, 6> calibrated_channel_data;
-					for (size_t j = 0; j < 6; j++)
-					{
-						raw_channel_data[j] = msg->raw_channel_data[loc].raw_channel_data[j];
-						calibrated_channel_data[j] = msg->calibrated_channel_data[loc].calibrated_channel_data[j];
-					}
-					data[i].value_.setRawChannelData(raw_channel_data);
-					data[i].value_.setCalibratedChannelData(calibrated_channel_data);
-					data[i].valid_ = true;
-				}
-				else
-				{
-					data.push_back(hardware_interface::as726x::AS726xState("", 0));
-				}
-			}
-
-			command_buffer_.writeFromNonRT(data);
-		}
-};
 } // namespace
