+++ zebROS_ws/src/state_listener/src/state_listener_controller.cpp
@@ -9,14 +9,5 @@ PLUGINLIB_EXPORT_CLASS(state_listener_controller::JointStateListenerController,
 PLUGINLIB_EXPORT_CLASS(state_listener_controller::JointModeListenerController,
 					   controller_interface::ControllerBase)
 
-PLUGINLIB_EXPORT_CLASS(state_listener_controller::PDPStateListenerController,
-					   controller_interface::ControllerBase)
-
-PLUGINLIB_EXPORT_CLASS(state_listener_controller::MatchStateListenerController,
-					   controller_interface::ControllerBase)
-
 PLUGINLIB_EXPORT_CLASS(state_listener_controller::IMUStateListenerController,
-					   controller_interface::ControllerBase)
-
-PLUGINLIB_EXPORT_CLASS(state_listener_controller::TalonStateListenerController,
 					   controller_interface::ControllerBase)
